[gcc] Cinematic: Make float to int conversion explicit

+++ src/cinematic/Cinematic.cpp
@@ -347,7 +347,7 @@ void DrawGrille(CinematicBitmap * bitmap, Color col, int fx, CinematicLight * li
 				}
 				
 				if(interp != 1.f) {
-					u8 iinterp = interp * (Color::Limits::max() / ColorLimits<float>::max());
+					u8 iinterp = u8(interp * (Color::Limits::max() / ColorLimits<float>::max()));
 					Color color = Color::fromRGBA(AllTLVertex[uvs->indvertex].color);
 					color.a = std::min(color.a, iinterp);
 					AllTLVertex[uvs->indvertex].color = color.toRGBA();
